feature: add `Redis#getPersistentId`
Since we removed the magic which automagically set the persistent id to the options, this method is to provide the persistent id in case we connected to redis using the persistent flag.

If the persistent ID was provided via config, the ID from the config is returned. Otherwise, we return the persistent ID from `ext-redis`.
In case `update` flag is provided, the persistent ID is always returned from `ext-redis`.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>